+++ src/io_redirect.h
@@ -1,7 +1,7 @@
 #ifndef IO_REDIRECT_H
 #define IO_REDIRECT_H
+#include <stddef.h>
 
-
-int handle_io_redirection(int *argc, char *argv[], int *saved_stdout, int *saved_stdin);
+int handle_io_redirection(size_t *argc, char *argv[], int *saved_stdout, int *saved_stdin);
 
 #endif
